+++ .eslintrc.cjs
@@ -43,11 +43,7 @@ module.exports = {
         'import/no-extraneous-dependencies': [
             'error',
             {
-                devDependencies: [
-                    '**/*.test.ts',
-                    '**/*.test.tsx',
-                    '**/*.config.js',
-                ],
+                devDependencies: ['**/*/test/*', '**/*.config.js'],
             },
         ],
         '@typescript-eslint/no-unused-vars': [